refactor: move all core components to widgets
This introduces the new separation of what is a core component (style +
structure) and what is a widget (behavior). The idea is to keep core
components very stable and generic over time, and instead using the
concept of "widges" to mean things that are fully formed and usable
directly in an application.

This stability means that it becomes easier and safer to build new
components by composing core components. It also provides a clear
distinction between what is a core component and what is a widget.

BREAKING CHANGE: All @dhis2/ui-core exports have been migrated to
@dhis2/ui-widgets.