refactor: move to more explicit final-form api
This aligns the toggle components Switch, Checkbox, and Radio to use the
same style of API. It makes it more explicit by following closer to what
FinalForm has as an API.

    <Field type="radio" component={RadioControl} name="" label="" value="" />
    <Field type="checkbox" component={CheckboxControl} name="" label="" value="" />
    <Field type="checkbox" component={SwitchControl} name="" label="" value="" />

By hinting to FinalForm that a component is intended as a checkbox or a
radio, we can avoid using special event handling
(createToggleChangeHandler) and align it with how we handle events for
the other event types.

We can also use custom elements in the same group of controls:

    <Field type="checkbox" component={CheckboxControl} name="group-one" label="cat" />
    <Field type="checkbox" component={CustomCheckboxControl} name="group-one" label="dog" />

Note that we get automatic boolean value handling if we omit the value
prop. If we include it, we get automatic handling for strings through
FinalForm.

This changes the Field component to a generic component that can be
re-used across all *Field components, instead of having to reimplement
the same composition everywhere.

BREAKING CHANGE: Field now provides a composition to provide all
necessary things for a *Field component.

BREAKING CHANGE: RadioGroup has been deleted.

BREAKING CHANGE: CheckboxGroup has been deleted.

BREAKING CHANGE: CheckboxGroupControl has been deleted.

BREAKING CHANGE: RadioGroupControl has been deleted.