fix: fix crash when an access string contains data settings
If the access string for a user/group contained data access settings
(aka the 2nd "rw" part in the string) the internal conversion to access
constants was returning the default value for publicAccess, which is
ACCESS_NONE.
This was causing a crash because the dropdown select for user/group does
not have the None option.

The parsing now uses a regex to only match the correct part of the
access string for metadata sharing settings.