refactor: remove context and z-index from Layer
Introduces a new Portal component that can be used as a standard React
component, either with a default or custom mount point.

BREAKING CHANGE: Removes the Layer.level prop and manual control of
z-index based rendering. The Layer component has a zIndex prop that can
be used to manually control the z-index if needed.

BREAKING CHANGE: Removes LayerContext and hooks for it, e.g.
useLayerContext.

BREAKING CHANGE: Nested Modals in JSX is now an unsupported technique.
Modals can be siblings, but not placed within each other.

If a modal needs to create another modal, then a wrapper component that
controls which of the sibling modals should open and when it should be
used. Modals placed as siblings will render on top of one another based
on the order in the DOM, so last sibling renders on top of the
preceeding ones.

Jira-ticket: LIBS-109