feat: introduce ui-elements package
Important: note that this deprecates ui-core but does not remove it. UI
Elements are also not included in the @dhis2/ui package nor in
Storybook.

@dhis2/ui-elements

Adds the UI Elements package with an incomplete set of tests that are
intended to ensure that all the Element components comply to the
requirements.

Implements the Button using the UI Elements principles from the README.
This is the proposal of how an UI Element should be structured.

UI Elements will contain more components than exist in the Design
System, as we will need to divide each component into a set of UI
Elements.

We expect to have many UI Elements as we need to ensure that all our
higher level components (and more) can be composed using the UI
Elements.

Name conventions

To provide better support for automated tools, wider arrange of
developer environments, and be predictable, the name conventions
changes.

PascalCasing and camelCasing are reserved for source code use. The file
system should be kebab-case.

In the future, conventions will be enforced through linting.

These are the conventions applied:

-   directories should always be kebab-case

-   a component is organized into a subfolder with it's name:
    src/{component}

-   component files should be kebab-case with the suffix:
    src/{component}/{component}.component.js

-   component styles should have the suffix:
    src/{component}/{component}.styles.js

-   components stories should have the suffix:
    src/{component}/{component}.stories.js

-   component test files should have the suffix:
    src/{component}/{component}.spec.js

-   in case of multiple specs, they can be in subdirectory for the
    component named:
    src/{component}/tests

    -   test files under tests/ do not need the component as prefix

-   component feature files go in a subdirectory of the component:
    src/{component}/features