docs(props): remove jsdoc typedefs for props
We prefer to use React doc strings to document props, as that is what
most of the tools in the React ecosystem expects, e.g. Storybook and
React docgen.

Having JSDoc typedefs in additon means that we have two keep two
documentation styles in sync with the prop changes.

A follow-up for some components may be required to add relevant React
docstrings to props where we used to have typedefs coverage.