docs(jsdoc): remove jsdoc strings
We align with the way react-docgen extracts information from our
components, and that means that we can strip out our JSDoc strings. In a
lot of places they are stale, and in most cases they follow a format
that is not necessary for us.

Instead we can use a simpler format:

    /**
     * General component description.
     */
     const Component = ({ foo, bar }) => (...)

     Component.propTypes = {
        /**
         * multi-line description
         * of prop "foo"
         */
        foo: PropTypes.number,

        /** single-line description of prop "bar" */
         bar: PropTypes.number,
     }

This should be enough for most cases and tools that build on
react-docgen.