feat(portal): introduce component
Introduces a new Portal component that can be used as a standard React
component, either with a default or custom mount point.

Nested Modals in JSX is now an unsupported technique. Modals can be
siblings, but not placed within each other.

If a modal needs to create another modal, then a wrapper component that
controls which of the sibling modals should open and when it should be
used. Modals placed as siblings will render on top of one another based
on the order in the DOM, so last sibling renders on top of the
preceeding ones.