fix(button): remove z-index from border style
Modifying the z-index on the border causes it to be shown above the
normal stacking context, which causes it to overlay on top of e.g.
modals and component covers.

Generally we should avoid modifying the z-index of components and allow
them to be automatically put into the relevant stacking context.