feat(modal): constrain fluidity to viewport
The fluid modal is allowed to stretch to the entire viewport save for
some space around it to ensure that it is still perceived as a modal by
the user.

Within the ModalContent, a container (e.g. div or Box) can be placed
with either a specific size, or a dynamic size, and the Modal will
resize automatically to accomodate for the content.

If the content is small, the Modal will contract to fit it.