test(button): use mount and workaround for popper
A shallow render is not enough for the DropdownComponent, so by using
mount instead we don't have to use the default object in the
controlledOnClick function.

That reveals a second problem that is documented on the react-popper
issue tracker[1].

Basically, we need to flush the micro-task queue for Promises to make
sure that the async functions in Popper have completed and the render is
actually completed.

We do this by using the trick:

    await act(async () => await null)

after rendering the component that has a nested Popper.

[1] https://github.com/popperjs/react-popper/issues/350