refactor(dropdown-button): single click handler
Merges controlledOnClick and toggleOpenState into one handler that does
the correct action based on if the component is controlled or
uncontrolled.

Moves the isControlled helper to a class property so it can be used in
the handler and the render function.

Makes the onClick prop optional in all cases, and simplifies the
controlled state to depend solely on the 'open' prop.

Changes the controlled DropdownButton story to be based on the open
prop without state management so that there is a clear difference in
behaviour across the uncontrolled/controlled stories.