feat(org unit tree): load children when expanding org unit to save bandwidth
BREAKING CHANGE: The payload and behavior of the `onChildrenLoaded` prop
have changed slightly! Previously the callback would be called for every
child that's been loaded. Now the callback will be called once with the
node that has the children, so fewer calls but with a payload with an
org unit one level above of what would be that case before this change.