feat(layer): add `onBackdropClick` prop (#1060)
The `Layer`'s backdrop is now a sibling element of the children container. This has helped preventing some unexpected behaviour caused by event bubbling, as described in [LIBS-192](https://dhis2.atlassian.net/browse/LIBS-192). Having the backdrop as a sibling is also common practice in UI libs.

While working on this, it came to light that the original prop-name `onClick` was probably not well-chosen, because developers might expect this to be a regular click event handler, including the regular event bubbling behaviour. This wasn't the case prior to this PR because we were already ignoring clicks originating in the children, so the prop-name was a poor choice from the start.

As a result we decided to add a new `onBackdropClick` prop and we have deprecated the `onClick` prop. The behaviour of both props is identical.